refactor: modularize tokenizers
