+++ Plugins.properties
@@ -54,46 +54,13 @@ filter=org.omegat.filters3.xml.android.AndroidFilter \
     org.omegat.filters3.xml.xmlspreadsheet.XMLSpreadsheetFilter \
     org.omegat.filters3.xml.wordpress.WordpressFilter \
     org.omegat.filters3.xml.scribus.ScribusFilter 
-tokenizer=org.omegat.tokenizer.DefaultTokenizer \
-    org.omegat.tokenizer.LuceneArabicTokenizer \
-    org.omegat.tokenizer.LuceneArmenianTokenizer \
-    org.omegat.tokenizer.LuceneBasqueTokenizer \
-    org.omegat.tokenizer.LuceneBrazilianTokenizer \
-    org.omegat.tokenizer.LuceneBulgarianTokenizer \
-    org.omegat.tokenizer.LuceneCatalanTokenizer \
-    org.omegat.tokenizer.LuceneCJKTokenizer \
-    org.omegat.tokenizer.LuceneCzechTokenizer \
-    org.omegat.tokenizer.LuceneDanishTokenizer \
-    org.omegat.tokenizer.LuceneDutchTokenizer \
-    org.omegat.tokenizer.LuceneFinnishTokenizer \
-    org.omegat.tokenizer.LuceneFrenchTokenizer \
-    org.omegat.tokenizer.LuceneGalicianTokenizer \
-    org.omegat.tokenizer.LuceneGermanTokenizer \
-    org.omegat.tokenizer.LuceneGreekTokenizer \
-    org.omegat.tokenizer.LuceneHindiTokenizer \
-    org.omegat.tokenizer.LuceneHungarianTokenizer \
-    org.omegat.tokenizer.LuceneIndonesianTokenizer \
-    org.omegat.tokenizer.LuceneIrishTokenizer \
-    org.omegat.tokenizer.LuceneItalianTokenizer \
-    org.omegat.tokenizer.LuceneJapaneseTokenizer \
-    org.omegat.tokenizer.LuceneLatvianTokenizer \
-    org.omegat.tokenizer.LuceneNorwegianTokenizer \
-    org.omegat.tokenizer.LucenePersianTokenizer \
-    org.omegat.tokenizer.LucenePolishTokenizer \
-    org.omegat.tokenizer.LucenePortugueseTokenizer \
-    org.omegat.tokenizer.LuceneRomanianTokenizer \
-    org.omegat.tokenizer.LuceneRussianTokenizer \
-    org.omegat.tokenizer.LuceneSmartChineseTokenizer \
-    org.omegat.tokenizer.LuceneSpanishTokenizer \
-    org.omegat.tokenizer.LuceneSwedishTokenizer \
-    org.omegat.tokenizer.LuceneThaiTokenizer \
-    org.omegat.tokenizer.LuceneTurkishTokenizer
+tokenizer=org.omegat.tokenizer.DefaultTokenizer
 plugin.desc.dictionary=Bundled dictionary files connectors. There are two bundled connectors. LingvoDSL, connector \
   \ for .dsl, .dsl.dz, .dsl.gz and StarDict, connector for .ifo, *.dict
 plugin.desc.misc=Bundled misc plugins
 plugin.desc.script=Script engine
 plugin.desc.guiutil=GUI extension
-plugin.desc.tokenizer=Bundled tokenizers plugin
+plugin.desc.tokenizer=Default tokenizer
 plugin.desc.filters2=Bundled filters for various file types
 plugin.desc.filters3=Bundled filters for simple (linear, unilingual) XML files includes LibreOffice (ODF) and DocBook
 plugin.desc.filters4=Bundled filters for complex XML files includes MS Office (OpenXML, with complex markups) and XLIFF (bilingual)
