+++ languagepacks/plugin.properties
@@ -1,0 +1,65 @@
+#
+#  OmegaT - Computer Assisted Translation (CAT) tool
+#           with fuzzy matching, translation memory, keyword search,
+#           glossaries, and translation leveraging into updated projects.
+#
+#  Copyright (C) 2023 Hiroshi Miura
+#                Home page: https://www.omegat.org/
+#                Support center: https://omegat.org/support
+#
+#  This file is part of OmegaT.
+#
+#  OmegaT is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  OmegaT is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+
+License= GNU Public License version 3 or later
+Author=OmegaT team
+Link=https://omegat.org
+Name=Tokenizers
+Category=tokenizer
+Description=tokenizers
+Classes=\
+    org.omegat.tokenizer.LuceneArabicTokenizer \
+    org.omegat.tokenizer.LuceneArmenianTokenizer \
+    org.omegat.tokenizer.LuceneBasqueTokenizer \
+    org.omegat.tokenizer.LuceneBrazilianTokenizer \
+    org.omegat.tokenizer.LuceneBulgarianTokenizer \
+    org.omegat.tokenizer.LuceneCatalanTokenizer \
+    org.omegat.tokenizer.LuceneCJKTokenizer \
+    org.omegat.tokenizer.LuceneCzechTokenizer \
+    org.omegat.tokenizer.LuceneDanishTokenizer \
+    org.omegat.tokenizer.LuceneDutchTokenizer \
+    org.omegat.tokenizer.LuceneFinnishTokenizer \
+    org.omegat.tokenizer.LuceneFrenchTokenizer \
+    org.omegat.tokenizer.LuceneGalicianTokenizer \
+    org.omegat.tokenizer.LuceneGermanTokenizer \
+    org.omegat.tokenizer.LuceneGreekTokenizer \
+    org.omegat.tokenizer.LuceneHindiTokenizer \
+    org.omegat.tokenizer.LuceneHungarianTokenizer \
+    org.omegat.tokenizer.LuceneIndonesianTokenizer \
+    org.omegat.tokenizer.LuceneIrishTokenizer \
+    org.omegat.tokenizer.LuceneItalianTokenizer \
+    org.omegat.tokenizer.LuceneJapaneseTokenizer \
+    org.omegat.tokenizer.LuceneLatvianTokenizer \
+    org.omegat.tokenizer.LuceneNorwegianTokenizer \
+    org.omegat.tokenizer.LucenePersianTokenizer \
+    org.omegat.tokenizer.LucenePolishTokenizer \
+    org.omegat.tokenizer.LucenePortugueseTokenizer \
+    org.omegat.tokenizer.LuceneRomanianTokenizer \
+    org.omegat.tokenizer.LuceneRussianTokenizer \
+    org.omegat.tokenizer.LuceneSmartChineseTokenizer \
+    org.omegat.tokenizer.LuceneSpanishTokenizer \
+    org.omegat.tokenizer.LuceneSwedishTokenizer \
+    org.omegat.tokenizer.LuceneThaiTokenizer \
+    org.omegat.tokenizer.LuceneTurkishTokenizer
