+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneArabicTokenizer.java
