+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneArmenianTokenizer.java
