+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneBulgarianTokenizer.java
