+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneCatalanTokenizer.java
