+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneDanishTokenizer.java
