+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneGalicianTokenizer.java
