+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneLatvianTokenizer.java
