+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneNorwegianTokenizer.java
