+++ languagepacks/src/main/java/org/omegat/tokenizer/LucenePersianTokenizer.java
