+++ languagepacks/src/main/java/org/omegat/tokenizer/LucenePortugueseTokenizer.java
