+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneSmartChineseTokenizer.java
