+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneSpanishTokenizer.java
