+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneSwedishTokenizer.java
