+++ languagepacks/src/main/java/org/omegat/tokenizer/LuceneThaiTokenizer.java
