+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'OmegaT'
-include("externalfinder", "machinetranslators", "tipoftheday")
+include("externalfinder", "machinetranslators", "tipoftheday", "languagepacks")
