+++ test/src/org/omegat/tokenizer/LuceneEnglishTokenizer.java
