+++ test/src/org/omegat/tokenizer/LuceneFrenchTokenizer.java
